Fix/avoid wrong listener deletion on secondary proxy ports
During the periodic bridge MDB check we'd wrongly delete a listener for
a specific group on any but the first proxy port.

This is because after finding it on the first proxy port we'd bail out.
However the same listener might be installed on other proxy ports, too.

Fixing the issue by continuing on other proxy ports.