Fix late last leave issue by blocking MLD queries instead
When the last listener for a proxied port leaves then no according MLD
report could be seen on the according proxied port. This is likely
because the report is delayed and we removed the dummy->proxy-port
redirection before that.

To fix this issue, instead of enabling/disabling the dummy->proxy-port
redirection enable/disable proxy-port->dummy redirection for MLD
queries.

This way any delayed leave (or early join) will always be redirected and
transmitted from the dummy interface.