Fix egress non-IPv6 packets being dropped on proxy ports
So far the egress TC filters we installed on a proxy port accidentally
blocked any non-IPv6 packet, leading to blocked ARP or IPv4 for
instance. Even though we never intended to touch ARP or IPv4 etc. in the
first place.

We need to not only allow egress IPv6 on a proxy port, if it wasn't MLD
reports, but we need to explicitly allow any packet then.