Use a bridge instead of a dummy interface
The bridge has a built-in querier feature we might want to use
to not only proxy MLD reports but also queries eventually.
Therefore switching from a dummy to bridge interface.

This should also make the tc rules a bit more intuitive, as the tc
redirections will now make brmldpX behave a bit more as if it
were a veth between brmldpbX and br0.
(And we might actually consider using a veth instead of a brmldpX dummy
interface in the future and replace tc with netfilter, if all needed
features of tc were part of netfilter eventually?)